# docs/api/openapiv1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "AzureBank.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Login",
        "description": "Authenticate user and receive JWT token.",
        "requestBody": {
          "description": "Login credentials",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfLoginResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          { }
        ]
      }
    },
    "/api/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Register",
        "description": "Register a new user account with initial bank account.",
        "requestBody": {
          "description": "Registration details",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfRegisterResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          { }
        ]
      }
    },
    "/api/auth/refresh": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Refresh access token",
        "description": "Exchange a refresh token for a fresh access + refresh token pair (rotation).",
        "requestBody": {
          "description": "The current refresh token",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfRefreshResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          { }
        ]
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Get current user",
        "description": "Get current authenticated user information.",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfUserResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - authentication failed or is missing. The body is a ProblemDetails whose errorCode names the reason (e.g. AUTH_TOKEN_MISSING, AUTH_TOKEN_INVALID).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - authenticated, but not permitted to reach this resource (errorCode: ACCESS_DENIED).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Logout",
        "description": "Logout and invalidate session.",
        "responses": {
          "200": {
            "description": "Success message",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - authentication failed or is missing. The body is a ProblemDetails whose errorCode names the reason (e.g. AUTH_TOKEN_MISSING, AUTH_TOKEN_INVALID).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - authenticated, but not permitted to reach this resource (errorCode: ACCESS_DENIED).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/pin": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Set or change PIN",
        "description": "Enrols a PIN, or changes an existing one. CHANGING requires `currentPin`; enrolling does\r\nnot, because the account password already gated getting here. A `currentPin` is verified\r\nwith the same attempt-limiting as every other PIN check (ADR-0010), so wrong values count\r\ntoward the lockout and a locked PIN cannot be replaced even by supplying the correct one.\r\nSee ADR-0040.",
        "requestBody": {
          "description": "PIN to set",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetPinRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SetPinRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SetPinRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - authentication failed or is missing. The body is a ProblemDetails whose errorCode names the reason (e.g. AUTH_TOKEN_MISSING, AUTH_TOKEN_INVALID).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - authenticated, but not permitted to reach this resource (errorCode: ACCESS_DENIED).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Business Rule Violation - the proof this change needs is missing: the password when enrolling a PIN (errorCode PASSWORD_REQUIRED), or the current PIN when changing one (errorCode PIN_REQUIRED).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "description": "A URI reference identifying the problem type"
                    },
                    "title": {
                      "type": "string",
                      "description": "A short, human-readable summary (e.g., 'Business Rule Violation')"
                    },
                    "status": {
                      "type": "integer",
                      "description": "The HTTP status code (422)"
                    },
                    "detail": {
                      "type": "string",
                      "description": "A human-readable explanation of the business rule violation"
                    },
                    "errorCode": {
                      "type": "string",
                      "description": "Machine-readable error code (e.g., 'INSUFFICIENT_FUNDS')"
                    },
                    "traceId": {
                      "type": "string",
                      "description": "Request trace identifier for debugging"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/pin/verify": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Verify PIN",
        "description": "Verify user's PIN for step-up authentication.",
        "requestBody": {
          "description": "PIN to verify",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyPinRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyPinRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyPinRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOfObject"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - authentication failed or is missing. The body is a ProblemDetails whose errorCode names the reason (e.g. AUTH_TOKEN_MISSING, AUTH_TOKEN_INVALID).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - authenticated, but not permitted to reach this resource (errorCode: ACCESS_DENIED).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "ApiResponseOfLoginResponse": {
        "type": "object",
        "properties": {
          "data": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/LoginResponse"
              }
            ]
          },
          "message": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "ApiResponseOfObject": {
        "type": "object",
        "properties": {
          "data": { },
          "message": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "ApiResponseOfRefreshResponse": {
        "type": "object",
        "properties": {
          "data": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/RefreshResponse"
              }
            ]
          },
          "message": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "ApiResponseOfRegisterResponse": {
        "type": "object",
        "properties": {
          "data": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/RegisterResponse"
              }
            ]
          },
          "message": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "ApiResponseOfUserResponse": {
        "type": "object",
        "properties": {
          "data": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/UserResponse"
              }
            ]
          },
          "message": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "LoginRequest": {
        "required": [
          "email",
          "password"
        ],
        "type": "object",
        "properties": {
          "email": {
            "maxLength": 255,
            "minLength": 1,
            "type": "string",
            "format": "email"
          },
          "password": {
            "maxLength": 128,
            "minLength": 8,
            "pattern": "^(?=.*[a-z])(?=.*[A-Z])(?=.*[0-9])(?=.*[^a-zA-Z0-9])[\\x20-\\x7E]{8,128}$",
            "type": "string",
            "description": "Password must contain at least one uppercase, one lowercase, and one digit."
          }
        }
      },
      "LoginResponse": {
        "required": [
          "token",
          "expiresAt",
          "user"
        ],
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "refreshToken": {
            "type": [
              "null",
              "string"
            ],
            "description": "Refresh token (plaintext, shown once) for rotating the short-lived access token via\r\nPOST /api/auth/refresh. In the BFF deployment this is captured server-side and never\r\nreaches the browser. Deliberately NOT `required`: a consumer deserializing this response\r\nacross the service boundary (the BFF) must degrade gracefully on its absence rather than\r\nhard-fail. Login issues one on every success (an issuance failure fails the login), so it\r\nis non-null in practice — the nullability is a boundary-robustness allowance, not a\r\nsignal that login may omit it."
          },
          "user": {
            "$ref": "#/components/schemas/UserLoginInfo"
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "errorCode": {
            "type": "string",
            "description": "Stable, machine-readable identifier for the failure — the value clients branch on. Absent only on validation failures, which carry an 'errors' dictionary instead."
          },
          "traceId": {
            "type": "string",
            "description": "Correlates this response with the server-side log entry. Quote it in a bug report."
          }
        }
      },
...
      "RefreshRequest": {
        "required": [
          "refreshToken"
        ],
        "type": "object",
        "properties": {
          "refreshToken": {
            "minLength": 1,
            "type": "string"
          }
        },
        "description": "Request body for POST /api/auth/refresh: exchanges a valid refresh token for a fresh\r\naccess + refresh token pair (rotation). The refresh token is the SOLE credential — no\r\nbearer access token is required (the access token being refreshed may already be expired)."
      },
      "RefreshResponse": {
        "required": [
          "accessToken",
          "refreshToken",
          "expiresAt"
        ],
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "Fresh JWT access token (short-lived, per JwtOptions.ExpirationMinutes)."
          },
          "refreshToken": {
            "type": "string",
            "description": "The NEW refresh token (plaintext, shown once). Store it; discard the old one."
          },
          "expiresAt": {
            "type": "string",
            "description": "Absolute expiry of the new access token (from its own exp claim).",
            "format": "date-time"
          }
        },
        "description": "Result of a successful refresh-token rotation: a new short-lived access token AND a NEW\r\nrefresh token. The presented refresh token is now revoked — the caller MUST replace its\r\nstored value with this one. Replaying the old token is rejected and, because it is now a\r\nrevoked token, trips reuse-detection (revoking the whole active token set for the user)."
      },
      "RegisterRequest": {
        "required": [
          "azureTag",
          "email",
          "password",
          "firstName",
          "lastName"
        ],
        "type": "object",
        "properties": {
          "azureTag": {
            "maxLength": 20,
            "minLength": 3,
            "pattern": "^[a-z][a-z0-9_]{2,19}$",
            "type": "string",
            "description": "Must start with a letter and contain only lowercase letters, numbers, and underscores."
          },
          "email": {
            "maxLength": 255,
            "minLength": 0,
            "type": "string",
            "format": "email"
          },
          "password": {
            "maxLength": 128,
            "minLength": 8,
            "pattern": "^(?=.*[a-z])(?=.*[A-Z])(?=.*[0-9])(?=.*[^a-zA-Z0-9])[\\x20-\\x7E]{8,128}$",
            "type": "string",
            "description": "Password must contain at least one uppercase, one lowercase, and one digit."
          },
          "firstName": {
            "maxLength": 50,
            "minLength": 2,
            "pattern": "^[a-zA-ZÀ-ÖØ-öø-ÿ\\s'-]{2,50}$",
            "type": "string"
          },
          "lastName": {
            "maxLength": 50,
            "minLength": 2,
            "pattern": "^[a-zA-ZÀ-ÖØ-öø-ÿ\\s'-]{2,50}$",
            "type": "string"
          }
        }
      },
      "RegisterResponse": {
        "required": [
          "user",
          "account",
          "token"
        ],
        "type": "object",
        "properties": {
          "user": {
            "description": "Registered user information",
            "$ref": "#/components/schemas/UserLoginInfo"
          },
          "account": {
            "description": "Initial primary account created for the user",
            "$ref": "#/components/schemas/AccountResponse"
          },
          "token": {
            "description": "Authentication token for immediate login",
            "$ref": "#/components/schemas/TokenResponse"
          }
        },
        "description": "Response returned after successful user registration.\r\nIncludes user info, initial account, and authentication token."
      },
      "SetPinRequest": {
        "required": [
          "pin"
        ],
        "type": "object",
        "properties": {
          "pin": {
            "maxLength": 6,
            "minLength": 6,
            "pattern": "^[0-9]{6}$",
            "type": "string",
            "description": "PIN must be exactly 6 digits."
          },
          "currentPin": {
            "maxLength": 6,
            "minLength": 6,
            "pattern": "^[0-9]{6}$",
            "type": [
              "null",
              "string"
            ],
            "description": "PIN must be exactly 6 digits."
          },
          "password": {
            "type": [
              "null",
              "string"
            ],
            "description": "The account password. Required when enrolling a PIN; ignored when changing an existing one,\r\nwhere `currentPin` is the proof instead."
          }
        }
      },
...
      "UserLoginInfo": {
        "required": [
          "id",
          "azureTag",
          "email",
          "firstName",
          "lastName",
          "hasPin"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "azureTag": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "hasPin": {
            "type": "boolean"
          }
        },
        "description": "User info returned after successful login (nested in LoginResponse)"
      },
      "UserResponse": {
        "required": [
          "userId",
          "azureTag",
          "email",
          "firstName",
          "lastName"
        ],
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "azureTag": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          }
        }
      },
      "VerifyPinRequest": {
        "required": [
          "pin"
        ],
        "type": "object",
        "properties": {
          "pin": {
            "maxLength": 6,
            "minLength": 6,
            "pattern": "^[0-9]{6}$",
            "type": "string",
            "description": "PIN must be exactly 6 digits."
          }
        }
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Enter your JWT token. Get a token by calling POST /api/auth/login",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ],
...
}